Improve cloud aggregation docs and support for smaller batches
Now buckets with fewer HTTP samples can also be aggregated, with outlier detection using the original sorting method, but with interpolation for improved accruacy. This is a more resource-intensive process, so for larger sample sizes quickselect is still the preferred algorithm.

There's also a new AggregationSkipOutlierDetection option that disables outlier detection altogether. It is disabled by default and is intended mostly for testing purposes.

And finally, the in-code documentation about the aggregation algorithm and options is expanded, which should help with https://github.com/loadimpact/k6/issues/628 and https://github.com/loadimpact/k6/issues/743