Equalize k6 js objects' fields names returned by setup
When goja bridges go objects it makes all Public fields to be case snake
as is the js standard. But because of lack of json tags when we marshall
and then unmarshall them due to them being returned by setup, the names
were becoming the go struct ones.
This was obviosly confusing and not intentional so now all objects that
are bridged by goja have json tags added to them to fix it.