Add a flag to disable the end-of-test summary
This, when used in conjunction with --no-thresholds, should reduce the k6 memory usage for long-running tests. This is something like an all-or-nothing approach to https://github.com/loadimpact/k6/issues/764 - even though it's not the best solution, it's pretty easy to implement and support and we can always extend it later.