Stop leaking VUs which were called once
This is due to them being referenced from a goroutine waiting for either
them to be interupted or the provided context to be Done. This works
okay in actual uses of k6 because each time the VU runs a script it
"interupts" the previous run (which has already finished so it does
nothing). In the tests though and especialy those provide
context.Background() and making VUs to just run a script once this is
not the case. There basically we leak each and every VU + everything in
points to.