k6 cloud: Correctly merge options.ext.loadimpact see c6c928a
After the fixes in c6c928a k6 basically unmarshalled the whole
options.ext.loadimpact in cloud.Config merged it with all other settings
and marshalled it back.
This would've worked perfectly if cloud.Config had everything the
options.ext.loadimpact could have as a field. But it doesn't and this is
by design, which leads to this fields disappearing.
The whole idea of options.ext is that it's something that will be parsed outside
k6 and will - k6 just copies it around.
But for three fields there is an exception k6 actually wants to merge
configs for projectID, token and name of the script. This obviosly has to be
refactored in the future to be outside the options.ext.loadimpact, but
currently this will be too big of change.

The new fix is that we will unmarshall in map[string]interface{} so that
even fields we don't know what to do will be saved and than "manually"
merge the three fields we want and then marshal the map back to json
keeping the unknown fields intact.