Exit with non zero code when setup/teardown timeouts fix #868
There are now 4 new non zero exit codes that will be revisioned at later
date when we more exit codes arise.

The 4 new ones are for each of setup and teardown timeouts, if it is a
timeout in the engine but it is not in setup/teardown - no such is known
at this time and 1 for engine error which is not timeout.