Refactor the http.request from the js part of k6 to lib.netext
I would've liked if I have moved more or ... less but this is the amount
of code that will be useful for future development and is the amount
that I could move with some amount of certainty that I am not totally
breaking the whole project.

There is the very real possibility that some types like URL and Response
might need to be copied because of further problems with goja but for
the moment this appears to be working.

We also loose the caching of the JSON representation of a response ...
which I don't know if it is actually useful or not ...