Add error_code tag to http metrics resolves #895
This also sets the error tag to the now more non dynamic messages.

This also changes a little bit of the behaviour specifically in that if
there is an error while making the http request that is not from the
RoundTripper than we consider this to be exception worthy now and
independantly of `throw` it will result in a exception. This should not
be ... common at all: the only error that is suppose to be able to
happen is if a redirect has an unparsable/missing location.