Support import cycles by returning the currently evaluated exports
This works by prepopulating what a given file(module) will export before
actually evaluationg the code whatsoever.

This way if it requires something that requires the original
module it will get the exports object as it is at that time. And because
of some magic (probably in babel) we also get that the exports are
bindings[1]. So even if something is not exported at the time of the
import, it will still be populated later when the original file finish
loading.

[1]: http://2ality.com/2015/07/es6-module-exports.html#why-export-bindings

Fixes #502