Add ability to compress request body (#989)
This adds `compression` param to all http methods with that will compress
 the body set the correct `Content-Encoding` and `Content-Length` and will 
send the request with the compressed body. The current compressions supported 
are `gzip` and `deflate` and their combinations separated by a comma (`,`).

If Content-Length of Content-Encoding has been set manually by the user,
they won't be reset but instead a warning will be printed.