Correctly handle exportes when module.exports get's redefined
This was broken in 3506ee1a, but as this mostly happens in browserified
code I didn't catch it as a problem. Now we get the "exports" once
before we run the script - to be used to resolve import cycles, and once
after the script is ran in order to get the exports if exports gets
redifined in the script.

This is probably broken in the case where both exports is redifined and
there is an import cycle, but the fix for this seems to be some magic
that will need to find out that we are actually requiring a new file and
then to "reexport" the exports before the import? or something like
that.