fix: Always initialize filesystems for all schemes always
This was always done when running a script, but for archives it was
deemed not needed as if it wasn't needed during the making of the
archive it shouldn't be needed during it's running. The problem arises
when a `require` is in a try block in order (to check if something is
support for example). In that case if it isn't loaded (for example a
nodejs module will not be loaded but will be resolved to an `https` url)
it will not be cached.

So if someone tries to run an archive from such a
script and there were no https imports k6 will panic at runtime.