Fix module.exports issues
The first issue was that if module.exports was set in the test directly,
for some reason it will than be changed when you get the cached version
of the `exports`. This seems like some goja specific thing that I
haven't been able to find out exactly why and where it happens.

The second was that `module.exports` or just `exports` couldn't be used
inside a function call(for example) to address an export of the module
you are in. This turned out to be due to the way module and exports
being set on the global object/scope in the previous code. This was done
before each import/require in order for subsequent exports/modules to
not overwrite previous once. But after that the old ones were set back
through the defers.

So that means that while during the execution of a module it sees it's
`exports/modules` it sees the ones of the module that is entrypoint for
the main function call. Usually(always in real k6 execution)  this means
the one of the main module - the one with the `default` function. Which
is exactly what happens in the test - the function call to `B` during
the execution of the main `default` function sees the `exports/modules`
for the main script.

When I changed it to instead have arguments to the function wrapper and
provide the `exports/module` this way, they shadow the global ones and
everything works as expected.

The main module will continue to use global `exports/module` as it is
set in a different place.