Rename schedulers to executors without renaming files or packages
This includes all of the renames announced in https://github.com/loadimpact/k6/pull/1007#issuecomment-497319205:
- schedulers are renamed as executors
- lib.ExecutorState is now lib.ExecutionState
- the lib.Executor interface and the local.Executor are now ExecutionSchedulers
- the `manual-execution` executor is renamed to `externally-controlled`
- incomplete iterations are now referred to as interrupted in the CLI UI

All of these changes are in the code and comments only, no files or packages have been moved or renamed. This will happen in a follow-up commit, so that the git move tracking doesn't break, due to all of the changes in the file contents.