cmd: make stats/status always output result
Currently, if no test is running, stats/status reports the error which
indicate that client can not connect to api server. From user
perspective, this message is not friendly.

Instead, we can return a more friendly message to user, and one still
can see the actual error when enable debug mode with "-v".