fix(httpext): return error on unsupported 101 response
The hanging behavior described in #971 happens after this Golang change:
https://github.com/golang/go/commit/541620409dee210c5498cc38433dcf690f58f888
, released in 1.12. If a `101 Switching Protocols` response is received,
`response.Body` will be replaced by an underlying `net.Conn`, and
reading it blocks indefinitely without a protocol implementation.

This is not a big problem for k6 in this case, since we don't support
protocol upgrades with the current HTTP API as shown in the example, so
the agreed solution was to detect it and return an error.

This Golang change is currently incompatible with client timeouts[1],
though this will hopefully be resolved by the time we add support for
h2c (#970) or other protocols.

[1]: https://github.com/golang/go/issues/31391

Closes #971