Prefix all envconfig struct tags with K6 directly
Works on #671 and fixing it in almost all cases.

This was done primarily by running:
```
rg -g "*.go"  'envconfig:"' --files-with-matches  |\
    xargs -n 1 sed -s -i 's/envconfig:"/envconfig:"K6_/g'
```

And than manually fixing all remaing problems by hand.
In the mean time I found out we weren't properly combing the
configuration for statsd/datadog and fixed.

Also renamed HttpDebug to HTTPDebug in the Options struct.

Unfortunately due to the way the statsd and datadog share the same
struct for a config this doesn't fix the issue as a whole for them.