test(executor): set lib.State for MiniRunnerVU, improve TestPerVUIterations
This models MiniRunnerVU closer to the way js.VU works, and is enough to
cleanup TestPerVUIterations to avoid abuse of logging.

There is no need to emit metrics in this case, since this is a lower
level baseline test, and for higher level functional tests involving
js.Runner we can use `__VU` or `__ITER` to track the result.