lib/netext/httpext: set request context explicitly
Go 1.13 and below has a bug that client timeout will not be propagated
to http request context. That bug was fixed already and will be present
in go1.14 (https://github.com/golang/go/issues/31657).

Currently, we set Timeout in our Client, but do not use it else where.
It will not affect the request context for now, but does when go1.14
release.

To make it compatible and works correctly with other go versions, we set
the request context explicitly with the timeout.

Update #1260