all: rename PushIfNotCancelled
The purpose of PushIfNotCancelled is pushing the sample if the passing
context is done. But its name will make confusion, because a timeout
context also causes the context cancelled.

So changing the name to PushIfNotDone to make it clearer. While at it,
also change the select/ctx.Done pattern to ctx.Err variant. It saves us
some lines of code and maybe slightly better performance.