Fix data race reading progress/logger of ExternallyControlled executor
Previously TestExternallyControlledRun would trigger a race condition
(when testing with `go test -race`) when writing the new config from one
goroutine and calling executor.GetProgress() or executor.GetLogger()
from another.

I'm not sure if using `configLock` for this is the right approach or if
there should be a separate mutex for this use case, so let me know.
I didn't benchmark if this impacts performance, but I wouldn't expect it
to as these should be read infrequently.