Rename ProgressBar.String -> ProgressBar.Render
It doesn't make sense to keep calling this String() since it breaks the
fmt.Stringer interface because of the padding argument, and more arguments
will need to be added to control the rendering.

I suppose we could still implement Stringer and have it call Render() with
some sane defaults, but I don't think we have a use case for that API,
and we can add it if needed.