Fix progress bar panic for empty segment
For an empty segment, we currently do nothing. But the goroutine spawned
to track progress bar still run, with and invalid progress information
passed to ProgressBar.

It causes the calculation in ProgressBar.String becomes invalid due to
overflow.

To fix this, we only do calculation when we have a valid filled value.

Fixes #1295