Fix externally controlled executor not pausing (#1479)
The problem was that pausing used gracefulStop which in turns wait for
the VU used by handle to be returned. The fix is to have another context
that is used for the actual runContext of the VU so we can cancel it
when we get gracefully stopped.

I tried to simplify (and speed up) the looping by only going through all
the channels and context stuff if there was a change, otherwise we use a
int32 and some atomic.LoadInt32 to see that "there is no change" and we
directly start a new iteration.

Additionally now the getVU error is propagetted by the vu handle's
start, which is probably badly handled in this commit.

Also the function that does the iteration now returns whether it was
interrupted mostly to help with faster interruption of the fast looping.
This might be reverted in a future commit, if the speed up isn't wanted.

Unfortunately in order to support returning the VU and not returning it
and still gettign the new VU but still to get start to actually return
the error from getVU it took way more code then was anticipated ...

The speed up is "insignficant" too, by my proffiling we were spending around
150-250ms (of 30s) in the mutex locking/channel reading now we spend 10ms(of
30s) in loading that int32. So 15-25x, but this is some sub percentage
time and this is for an empty bodied `default` function. More benchmarks
to come :D.

There is also a not so great benchmark :D

name                  iterations/ns
VUHandleIterations-8      0.04 ± 1%

so around 40ns per iteration