rewrite vuHandle as a state machine and fix some more races (#1506)
this is ... mostly a complete rewrite of vuHandle now using an atomically updated state to work as a state machine. 

There is still a fast path which provides all the speed, but now it also always returns any VU it has got and doesn't return it in cases where this isn't required - there was a start after a gracefulStop, but before it actually stopped. 

closes #1492