Replace BenchmarkCloud with just and aggregation benchmark
The previous benchmark was mostly a try at benchmarking aggregation, but
it was bad in ... a lot of ways. The biggest of which was that it didn't
actually wait for the aggregation to happen so it was mostly measuring
how fast Collect works, which is not ... as interesting.

Other problems were that it didn't have enough variation based on the
one thing that slows down aggregation - different tags, which was also
added.

Also removing the pushing of the metrics from the equation means that it
is less noisy.