Fix wrong name tag for HTTP metrics from redirects
Closes #1474

The root cause was reusing the same 'name' tag for all requests
processed by a transport. In the case of redirects and digest auth the
transport would process more than one request and needs to update the
'name' tag in the same way as 'url' is, but only for the initial
request, as subsequent ones will by dynamically generated.