Reuse context.Canceled for when we interupt a VU
This is actually what should've been to begin with ... as k6 is trying
to signal to the goja.VM that it should stop because the context was
canceled (or the deadline was reached).

Previous to this there was the chance that the goja.VM will be in a call
that will throw the context.Canceled on it's own (sleep) if the context
was canceled.

Fixes #1485