Add CompilerPool for tc39 tests where it was bottleneck
A lot of tc39 test need to compile and transpile the code of the test
before they can execute it. Babel *does* take significant amount of time
to transpile even small programs and having only one means that on
machines with 8 cores most of the time is being spent waiting for babel.

This can be used in other places at a later point - `k6/http` and some
`js` package libs are also mostly babel transpile dominated.

This speeds up the running of the test by around 4x on my machine. An