Run babel on compilation errors as well as on parsing ones
Unfortunately parsing checks syntax, not whether what the syntax did was
actually ... supported(sometimes).

As can be seen this fixes a lot of cases where this is a problem.

This is actually probably going to be an even bigger problem in the
future as goja gets more features in it which will let the parsing
to pass even though some feature actually isn't supported but the syntax
is now "valid".

Good example (which was how this was kind of found) is that destructuring
looks a lot like an object with shorthand properties, but this is only
noticable once the code is compiled, not while parsing. So in the future
when goja for example has support for all the rest of the syntax (like
`let`) it will be harder to catch.

Even now code such as `let [x] = [2]` which is desctructuring doesn't
return an error on the parsing stage but instead in the compilation.
This example might be a bug in goja, but I would argue we can have
others.