Maintenance update of github.com/spf13/cobra to v1.1.3 from v0.0.4+
The majority of the changes a dependency bumps and fixes for different
shell completions which k6 doesn't use.

The only added feature that will likely be used is adding
context.Context to cobra.Command which can be used in refactoring `cmd`
further and dropping even more global variables.

Notable dependency updates:
- github.com/spf13/afero just added go.mod
- github.com/pkg/errors - nothing interesting
- github.com/mitchellh/mapstructure - this gets update from pre v1.0.0
  and has a ton of changes. We likely don't use it enough but still
  could be a problem
- github.com/spf13/pflag - mostly additions that might be used by k6
- gopkg.in/yaml.v2 - some change around really *long* string values that
  I don't understand but is actually reverting to pre v2.3.0 behaviour
  so hopefully doesn't break anything silently for us