When adding a VU to the pool wait for it the goroutine to starting
This helps with the fact that there is no gurantee currently that just
because we have added 100 VUs any of them had actually started wait for
an iteration.

This likely has some downside in creating more channels when adding VUs
back and forth but it should be minimal compared to the amount of work
needed to call a JS function.

This additionally could've meant that for high rate runs it was possible
for not all VUs to be available leading to k6 not actually being able to
run iterations at the given rate even though it should've been able if
all the VUs had actually started.