Use a single segmented index for local and global iterations (#2057)
* Use a single segmented index for local and global iterations

This removes the need for a lot of the additional synchronization.

This also makes it so there are global iterations for all executors. For
some of those it means that the global iteration will be *unique* among
instances but doesn't necessary mean that iterations with smaller
numbers *have* run or *will* run in the same test run.

This is somewhat like how it is for the executors for which we already
had it, there is nothing making certain that a shared-iterations
executor will not run out of time to run all the iterations on one of
the instances or a arrival-rate to not have to drop iterations due to
lack VUs to execute them.

* Update lib/executor/base_executor.go

Co-authored-by: Ivan Mirić <ivan@loadimpact.com>

* unexport and comment nextIterationCounters

* Remove synchronization in SegmetnedIndex

* Drop SegmentedIndexResult

* Fix race in TestSharedIterationsGlobalIters

There is nothing stopping 5 VUs starting a shared iterations at the same
time and one of them getting "slower" to the actual execution than it's
iteration counter suggests.

Co-authored-by: Ivan Mirić <ivan@loadimpact.com>