Fix races due to single module instance
While previously common.Bind will wrap the one instance of the module
object in *separate* wrappers. After the previous change we no longer
wrap them so setting the getContext function for them just resets it
each time it gets imported, which even more importantly means that the
last VU which imports is the one which context will be used by
everything.

I specifically made this break if you implement the method but hasn't
implemented the per VU module instance interface so it can actually
break when a developer gets it wrong.