Fix CI process to only depend on tests with latest stable Go
We don't test with Go 1.15 at the moment because if using go:embed. The test-prev check is still here, but it is disabled and nothing depends on it. This will be the state until Go 1.17 is released and we can switch test-prev to 1.16. And even then, the build probably shouldn't depend on test-prev, its failure should be only informative, not aborting the whole CI.