Fix remaining gracefulSteps bug
handleVUs and handleRemainingVUs was not sharing the current number of
planned VUs in graceful steps. This may cause the latter method to act
strangely.

It's because handleRemainingVUs needs to know the number of
remaining graceful steps so that it can stop the remaining VUs
accordinly.

Now the handleVUs method returns the number of handled graceful steps.
So the handleRemainingVUs method can use this information and hard stop
the remaining VUs.