Make ChangePathFs compatible with new afero.Fs without updating
Given that we won't be using this method I don't think we need to have
actual real implementation for it which will require us to update
aforo.Fs, so I just returned an error.

This is predominantly motivated by the fact that an extension having a
dependency on a newer afero.Fs ( through another dependency for example)
will not compile as the interface afero.Fs was extended in v1.5.0.