Base main extension context on the context received in launch()
This is the context passed from k6, and for synchronization purposes it
would be better if the extension depended on it, so that errors like the
`panic: send on close channel` from issue 49 wouldn't happen. This
technically probably doesn't fix that yet, since the context created
in k6 isn't synchronized with the Engine shutdown and closing of the
SampleContainer channel, but it seems like a good idea regardless.