Fix nil pointer panic on Frame.networkIdleCtx
For very quick requests there's a race condition between when we set
`f.networkIdleCtx` to nil[1] as part of the `FrameManager.requestStarted()`
call, and when we start the idle timer again in `FrameManager.requestFinished()`[2].

This fix ensures we keep the reference to `networkIdleCtx` as part of the closure
in the goroutine crated in `Frame.startNetworkIdleTimer()`, and avoid
accessing `f.networkIdleCtx` directly.

This could use better synchronization, but this fixed it consistently for
me by testing with `examples/getattribute.js`, which also reproduces the
issue (see https://github.com/grafana/xk6-browser/runs/4225921786).

[1]: https://github.com/grafana/xk6-browser/blob/baaf58caef73370f1659c30b3d3a989ddcd8da27/common/frame.go#L273

[2]: https://github.com/grafana/xk6-browser/blob/baaf58caef73370f1659c30b3d3a989ddcd8da27/common/frame.go#L290

Fixes #109