Implement RemoteObject conversion to native types
Previously we used to throw an error if we encountered a custom object
as returned by CDP from an `Evaluate()` or `console.log()` call.

This should handle those cases properly now by converting objects to
`map[string]interface{}` internally, and using a custom Logrus formatter
to marshall the map to JSON for safe output when used with `console.log()`.

Fixes #120