Fix interface conversion in document
The error was:
panic: interface conversion: interface {} is nil, not
*common.ElementHandle

The cause of the error was trying to assert a nil result
in Frame.document to *ElementHandle. Of course, this is
probably the surface error. There may be other causes
for it, or maybe not.

It was hard to test Frame, so I used an abstraction called
FrameExecutionContext. There are a lot of problems are
happening with execution contexts. See: #125 and #49, for
example. So, it's better to make them testable, starting
here.

Fixes #53