Fix cannot create frame session
We cannot create a FrameSession and context naturally gets
canceled When the following error occurs:
"cannot find context with specified id"

So the fix is about gracefuly handling this situation and
preventing the context cancelation error propagating into
the FrameSession.

There were also missing locks for execution contexts.

This also adds for a simpler, locking-based detection
for execution context changes. The previous one was
error prone because atomics better not be used for
race condition issues.