Fix (#49) cannot find context with specified id
Sometimes sessions get closed and they clog
the communication between frame sessions and
sessions. This usually happens when a frame
session is getting created after new frame
attachments. This was causing problems with
frame handling.

This commit prevents lagging by failing
fast when it detects a session is closed.

It also changes the ordering of FrameSession
init steps. The reason of putting initOptions
last is: It was propagating frame events
before the frame is ready to be used.