Update of the golangci-lint version & refactor of make check
Updated the golangci-lint to the v1.43.0, cleaned the configuration file.

Small refactoring of Makefile. Introduced a new Makefile target, `make ci-like-lint` to run the linter inside the container with the same version used in CI (run the specific version in the docker). The `make check` target uses `make ci-like-lint` by default.

Closes #2294