Fix WaitForFrameNavigation nil pointer bug
`event` in `WaitForFrameNavigation` becomes nil when context is canceled
, hence the nil pointer error when trying to use the event's fields.

Browser was throwing an error when the context is canceled, but it
shouldn't do so if the extension is in the shutting down stage. These
problems were popped up while trying to test the nil pointer error.

This is a critical method for creating the browser pages where all the
action is happening. So I wanted to add a unit test for this as well.
This way, I can test it from all angles (hopefully).

browser.go:
This commit removes the conn field from the BrowserContext because it
never actually uses it. It's better to localize value owners. In this
case, the Browser is the actual owner of the connection.

We incorrectly named connMu because it doesn't protect the conn field,
instead, it protects the connected state. So this commit changes its
name from connMu to connectedMu.

Added a new function named newBrowser that can return a Browser value
without an actual connection to a running browser. This helped me to
unit test the Browser without actually connecting to a running Browser.